Add various style assets, DEVELOPERS guide
